pyproject.toml
[tool.poetry]
name = "dsspy"
version = "1.0.0"
description = "A Python implementation of the DSSP algorithm for secondary structure assignment of proteins."
authors = ["Robert Fenwick"]
readme = "README.md"
packages = [{include = "dsspy"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
biopython = "^1.83"
numpy = "^2.3.2"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
pylint = "^3.1.0"
mypy = "^1.9.0"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
deptry = "^0.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.deptry]
extend_exclude = ["test"]

[tool.poetry.scripts]
dsspy = "dsspy.cli:main"
